style: 부등호 방향, 변수 타입, import 변경

+++ src/main/java/org/project/paysystem/entity/GlobalPricing.java
@@ -18,7 +18,7 @@ public class GlobalPricing {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int unitPrice;  // 광고 단가 (금액)
+    private double unitPrice;  // 광고 단가 (금액)
 
     @Column(nullable = false)
     @Enumerated(value = EnumType.STRING)
